refactor: move the audited transfer flow from the ring cli into the ring sdk
