+++ custom-rings/cli/Cargo.toml
@@ -31,15 +31,11 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 solana-address = { workspace = true }
 solana-loader-v3-interface = { workspace = true, features = ["serde"] }
-solana-address-lookup-table-interface = { workspace = true }
-solana-compute-budget-interface = { workspace = true }
 solana-instruction = { workspace = true }
 solana-keypair = { workspace = true }
-solana-message = { workspace = true }
 solana-signature = { workspace = true }
 solana-signer = { workspace = true }
 solana-system-interface = { workspace = true }
-solana-transaction = { workspace = true, features = ["serde"] }
 solana-transaction-error = { workspace = true }
 toml = { workspace = true }
 zolana-client = { workspace = true, features = ["indexer-api", "solana-rpc"] }
